[refactor] Refactor match scheduling and venue utilities
