+++ Frontend/utils/date.ts
@@ -6,3 +6,10 @@ export function isToday(date: Date): boolean {
     date.getDate() === today.getDate()
   );
 }
+
+export function parseDraftDate(value?: string): Date | undefined {
+  if (!value) return undefined;
+  const parsed = new Date(value);
+  if (Number.isNaN(parsed.getTime())) return undefined;
+  return parsed;
+}
